+++ testutils/src/main/java/com/google/android/exoplayer2/testutil/TestDownloadManagerListener.java
@@ -100,7 +100,8 @@ public final class TestDownloadManagerListener implements DownloadManager.Listen
   }
 
   @Override
-  public void onDownloadChanged(DownloadManager downloadManager, Download download) {
+  public void onDownloadChanged(
+      DownloadManager downloadManager, Download download, @Nullable Exception finalException) {
     if (download.state == Download.STATE_FAILED) {
       failureReason = download.failureReason;
     }