+++ RELEASENOTES.md
@@ -149,6 +149,8 @@
         initialization vector used to encrypt the cache contents.
     *   Add `Requirements.DEVICE_STORAGE_NOT_LOW`, which can be specified as a
         requirement to a `DownloadManager` for it to proceed with downloading.
+    *   For failed downloads, propagate the `Exception` that caused the failure
+        to `DownloadManager.Listener.onDownloadChanged`.
 *   Audio:
     *   Add a sample count parameter to `MediaCodecRenderer.processOutputBuffer`
         and `AudioSink.handleBuffer` to allow batching multiple encoded frames